+++ HISTORY.rst
@@ -6,7 +6,13 @@ History
 0.3.1 (unreleased)
 ------------------
 
-* No changes yet.
+* Fix issues #108 #109 and merges #110 #120
+* Remove requirements.txt, add docs/rtd-pip-requirements, update .travis.yaml
+* Add support for running at sites where data is locally visible (e.g., nersc)
+* Add support for reading spArc and spFlat calibration files.
+* Add CalibrationTutorial notebook.
+* Add bosssky script.
+* Fix some issues with python3 and astropy < 2.
 
 0.3.0 (2017-12-17)
 ------------------